+++ src/orderbook/OrderBook.h
@@ -14,43 +14,43 @@ class OrderBook {
     /**
      * Represents bids, descending in nature.
      */
-    std::array<double, ORDERBOOK_MAX_DEPTH> bid_price;
-    std::array<double, ORDERBOOK_MAX_DEPTH> bid_qty;
+    std::array<float, ORDERBOOK_MAX_DEPTH> bid_price;
+    std::array<float, ORDERBOOK_MAX_DEPTH> bid_qty;
 
     /**
      * Represents asks, ascending in nature.
      */
-    std::array<double, ORDERBOOK_MAX_DEPTH> ask_price;
-    std::array<double, ORDERBOOK_MAX_DEPTH> ask_qty;
+    std::array<float, ORDERBOOK_MAX_DEPTH> ask_price;
+    std::array<float, ORDERBOOK_MAX_DEPTH> ask_qty;
 
     int bid_size;
     int ask_size;
 
-    static void update_side(std::array<double, 10> &side_price, std::array<double, 10> &side_qty, int &side_size, double price, double quantity, bool
+    static void update_side(std::array<float, 10> &side_price, std::array<float, 10> &side_qty, int &side_size, float price, float quantity, bool
                             is_ask);
 
     /**
      * Checks if update can be inserted in the first place.
      */
-    static bool is_updated_not_needed(const std::array<double, 10> &side, int side_size, double price, double quantity, bool is_ask);
+    static bool is_updated_not_needed(const std::array<float, 10> &side, int side_size, float price, float quantity, bool is_ask);
 
     /**
      * Returns the position of the target price or the index, which is the smallest price greater than (ask) or biggest price smaller than (bid).
      * This is the position to apply updates to.
      */
-    static int search_side(double target_price, int side_size, const std::array<double, ORDERBOOK_MAX_DEPTH>& side_price, bool is_ask);
-    static int search_side_linear(double target_price, int side_size, const std::array<double, ORDERBOOK_MAX_DEPTH>& side_price, bool is_ask);
+    static int search_side(float target_price, int side_size, const std::array<float, ORDERBOOK_MAX_DEPTH>& side_price, bool is_ask);
+    static int search_side_linear(float target_price, int side_size, const std::array<float, ORDERBOOK_MAX_DEPTH>& side_price, bool is_ask);
 
     /**
      * Updates the order book if incoming price is already a level on the book.
      */
-    static void found_price_update_side(int index_to_update, std::array<double, 10> &side_price, std::array<double, 10> &side_qty, int &side_size, double
+    static void found_price_update_side(int index_to_update, std::array<float, 10> &side_price, std::array<float, 10> &side_qty, int &side_size, float
                                         quantity, bool is_ask);
 
     /**
      * Updates the order book if incoming price does not exist on the book.
      */
-    static void not_found_price_update_side(int index_to_update, std::array<double, 10> &side_price, std::array<double, 10> &side_qty, int &side_size, double price, double
+    static void not_found_price_update_side(int index_to_update, std::array<float, 10> &side_price, std::array<float, 10> &side_qty, int &side_size, float price, float
                                             quantity, bool is_ask);
 
 public:
@@ -60,8 +60,8 @@ public:
     /**
      * Update functions for the order book.
      */
-    void update_bid(double price, double quantity);
-    void update_ask(double price, double quantity);
+    void update_bid(float price, float quantity);
+    void update_ask(float price, float quantity);
 
     /**
      * Display functions
@@ -72,12 +72,12 @@ public:
     /**
      * Getters for book-level data like sizes & L1
      */
-    double get_bid_size() const;
-    double get_ask_size() const;
-    double get_spread() const;
-    double get_mid_price() const;
-    double get_top_bid() const;
-    double get_top_ask() const;
+    float get_bid_size() const;
+    float get_ask_size() const;
+    float get_spread() const;
+    float get_mid_price() const;
+    float get_top_bid() const;
+    float get_top_ask() const;
 };
